Add file for App component
